Added evalution functionality in the app.
