+++ lib/Screen/Pages/Registration/widget/studentRegInfo.dart
